KAIZEN-0 Refaktorering av heroku
Fjernet pre/post install hooks og la til rbedre URL håndtering for heroku server
