--- scripts/preHeroku.js
@@ -1,6 +1,0 @@
-const fs = require('fs');
-
-const pkg = JSON.parse(fs.readFileSync('./package.json', 'utf-8'));
-pkg.homepage = '/';
-
-fs.writeFileSync('./package.json', JSON.stringify(pkg, null, 2), 'utf-8');
